pyproject.toml
[project]
name = "20-chunking-semantic"
version = "0.1.0"
description = ""
authors = [
    {name = "RichardAyalaPH",email = "richard@purrfecthire.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "langchain-huggingface (>=0.3.1,<0.4.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "pypdf (>=6.0.0,<7.0.0)",
    "supabase (>=2.18.1,<3.0.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)"
]

[tool.poetry]
packages = [{include = "20_chunking_semantic", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
